Normalize Creature doubles and ints to floats

+++ core/src/main/java/com/bouncefish/utils/GameConstants.java
@@ -9,12 +9,12 @@ public final class GameConstants {
     private GameConstants() { } //We never want to instantiate this -- think like C# static class
 
     // If we want to make any of these editable at runtime, then we'll have to convert this class into something not 'final'
-    public static final double GRAVITY = 3000;       // Persistent downward acceleration
-    public static final double BOUNCE_DAMPING = 1; // Energy retained on bounce (maybe can make this creature specific?)
+    public static final float GRAVITY = 3000;       // Persistent downward acceleration
+    public static final float BOUNCE_DAMPING = 1; // Energy retained on bounce (maybe can make this creature specific?)
     public static final double DEATH_BOUNCE_DAMPING = 0.3; // Purely visual damping multiplier for when the player dies
     public static final double FRICTION = 0.98;      // Friction when bouncing on a creature (maybe can make this creature specific?)
     //public static final double H_ACCEL = 45;        // How fast bouncefish should speed up when inputting left/right (convert this to immediate velocity?)
-    public static final double H_SPEED = 800; // Base value for horizontal 'movement' speed
+    public static final float H_SPEED = 800; // Base value for horizontal 'movement' speed
     public static final double MAX_H_SPEED = 700;    // Maximum horizontal speed
     public static final float GROUND_HEIGHT = 0; // Y position of ground level
     public static final int RIGHT_CONTROL_BORDER = 400;
